Redo combo parsing
Instead of logging the position, log the keys, this way if the
keymap changes, we are still able to find the correct combo
by the key positions.

Next to that fixed quite some issues where shifted-combo's
would not be registered as shifted. Holds were not registered
either. But all of that is handled now.